fixing the css font family name in the previews to handle font names with spaces
