+++ packages/edit-site/src/components/global-styles/font-library-modal/font-demo.js
@@ -8,22 +8,22 @@ import { useContext, useEffect, useState, useRef } from '@wordpress/element';
  * Internal dependencies
  */
 import { FontLibraryContext } from './context';
+import { getFacePreviewStyle } from './utils/preview-styles';
 
 function FontFaceDemo( { fontFace, text, style = {} } ) {
 	const ref = useRef( null );
 	const [ isIntersecting, setIsIntersecting ] = useState( false );
 	const [ isAssetLoaded, setIsAssetLoaded ] = useState( false );
 	const { loadFontFaceAsset } = useContext( FontLibraryContext );
-	const { fontFamily, fontStyle, fontWeight } = fontFace;
+
+	const faceStyles = getFacePreviewStyle( fontFace );
 
 	const demoStyle = {
-		fontWeight,
-		fontStyle,
-		fontFamily,
 		flexShrink: 0,
 		fontSize: '18px',
 		opacity: isAssetLoaded ? '1' : '0',
 		transition: 'opacity 0.3s ease-in-out',
+		...faceStyles,
 		...style,
 	};
