+++ packages/edit-site/src/components/global-styles/font-library-modal/utils/index.js
@@ -2,6 +2,7 @@
  * Internal dependencies
  */
 import { FONT_WEIGHTS, FONT_STYLES } from './constants';
+import { formatFontFamily } from './preview-styles';
 
 export function setUIValuesNeeded( font, extraValues = {} ) {
 	if ( ! font.name && ( font.fontFamily || font.slug ) ) {
@@ -84,10 +85,14 @@ export async function loadFontFaceInBrowser( fontFace, source, addTo = 'all' ) {
 	}
 
 	// eslint-disable-next-line no-undef
-	const newFont = new FontFace( fontFace.fontFamily, dataSource, {
-		style: fontFace.fontStyle,
-		weight: fontFace.fontWeight,
-	} );
+	const newFont = new FontFace(
+		formatFontFamily( fontFace.fontFamily ),
+		dataSource,
+		{
+			style: fontFace.fontStyle,
+			weight: fontFace.fontWeight,
+		}
+	);
 
 	const loadedFace = await newFont.load();
 
@@ -118,71 +123,6 @@ export function getDisplaySrcFromFontFace( input, urlPrefix ) {
 		src = encodeURI( src );
 	}
 	return src;
-}
-
-function findNearest( input, numbers ) {
-	// If the numbers array is empty, return null
-	if ( numbers.length === 0 ) {
-		return null;
-	}
-	// Sort the array based on the absolute difference with the input
-	numbers.sort( ( a, b ) => Math.abs( input - a ) - Math.abs( input - b ) );
-	// Return the first element (which will be the nearest) from the sorted array
-	return numbers[ 0 ];
-}
-
-function extractFontWeights( fontFaces ) {
-	const result = [];
-
-	fontFaces.forEach( ( face ) => {
-		const weights = String( face.fontWeight ).split( ' ' );
-
-		if ( weights.length === 2 ) {
-			const start = parseInt( weights[ 0 ] );
-			const end = parseInt( weights[ 1 ] );
-
-			for ( let i = start; i <= end; i += 100 ) {
-				result.push( i );
-			}
-		} else if ( weights.length === 1 ) {
-			result.push( parseInt( weights[ 0 ] ) );
-		}
-	} );
-
-	return result;
-}
-
-export function getPreviewStyle( family ) {
-	const style = { fontFamily: family.fontFamily };
-
-	if ( ! Array.isArray( family.fontFace ) ) {
-		style.fontWeight = '400';
-		style.fontStyle = 'normal';
-		return style;
-	}
-
-	if ( family.fontFace ) {
-		//get all the font faces with normal style
-		const normalFaces = family.fontFace.filter(
-			( face ) => face.fontStyle.toLowerCase() === 'normal'
-		);
-		if ( normalFaces.length > 0 ) {
-			style.fontStyle = 'normal';
-			const normalWeights = extractFontWeights( normalFaces );
-			const nearestWeight = findNearest( 400, normalWeights );
-			style.fontWeight = String( nearestWeight ) || '400';
-		} else {
-			style.fontStyle =
-				( family.fontFace.length && family.fontFace[ 0 ].fontStyle ) ||
-				'normal';
-			style.fontWeight =
-				( family.fontFace.length &&
-					String( family.fontFace[ 0 ].fontWeight ) ) ||
-				'400';
-		}
-	}
-
-	return style;
 }
 
 export function makeFormDataFromFontFamilies( fontFamilies ) {
